fix: angka pembayaran tidak lagi tampil notasi saintifik (1.408E7 -> 14080000) | bump v2.5.0

+++ app/src/main/java/com/posbah/app/ui/screens/bmp/invoices/InvoicesViewModels.kt
@@ -239,7 +239,7 @@ class InvoiceDetailViewModel @Inject constructor(
     }
 
     fun selectEditPayment(payment: BmpInvoicePaymentEntity) =
-        _ui.update { it.copy(editingPayment = payment, newPaymentAmount = payment.paymentAmount.toString(), newPaymentMethod = payment.paymentMethod) }
+        _ui.update { it.copy(editingPayment = payment, newPaymentAmount = payment.paymentAmount.toLong().toString(), newPaymentMethod = payment.paymentMethod) }
 
     fun cancelEditPayment() =
         _ui.update { it.copy(editingPayment = null, newPaymentAmount = "", newPaymentMethod = "TRANSFER") }
